dist: Fix character encodings in NEWS
The NEWS file had a mix of ISO-8859-1 and UTF-8 encodings, which
was making a mess of decoding the non-ASCII characters in the
file.  This patch unifies the NEWS file as a UTF-8 encoded file
and changes many of the places where we had ASCII-ified a persons
name.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit 2e23893f04ef30c598a23889a4242f8cf4a45238)

Cherry-pick was modified to fix a few more ISO-8859-1 characters that
were in the v3.0.x branch but not in the master branch.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>