libevent: Upgrade Libevent to 2.1.12-stable
The refactoring patches move Libevent from a framework integration
to a 3rd-party package, but did not change the Libevent version
that Open MPI ships.  During that swap, we stopped running the
Autotools on Libevent and relied on the tools the Libevent authors
used when building the 2.0.22 release tarball.  The config.guess
in this release tarball did not work on the IBM systems.

This patch updates the release version of Libevent to 2.1.12-stable,
which will suck in a bunch of upstream bug fixes and updates
the config.guess so that the 3rd-party refactoring actually
compiles on the IBM Power systems.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>