orterun.1in: define "slot" and "processor element"
Add descriptive definitions of "slot" and "processor element" at the
top of the man page (and effectively delete / move some text from
lower in the man page up into those definitions).

Also add a little blurb in the --use-hwthread-cpus description about how
it changes the definition of "processor element".

This is not a cherry-pick from master because PRRTE has replaced ORTE
on master, and orterun.1in no longer exists in master.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>