flatten fixes (count vs flatten divergence)
This bug is hit by the same code I used long ago, darray_bug_romio.c.
I'm not totally sure on the archeology of when the bug reappeared,
but I think it was because of changes to COMBINER_RESIZED in
ADIOI_Flatten that weren't mirrored in ADIOI_Count_contiguous_blocks.

Count and Flatten are very similar functions, but they diverge a lot
around the edges and in degenerate cases.

The two changes I made are

1. A book-keeping rule in Flatten is *curr_index should be updated
in sync with whatever is being added to flat->indices[] and
flat->blocklens[].  There were a couple places (when counts[] arrays
have 0s) where that wasn't the case (see the "nonzeroth" var in this
commit).

2. The main change I made was to reset flat->count based on
Flatten's curr_index.  This is because the divergence between the
two functions is too great to reliably fix.

3. A third change is just a safety valve, using a
FIX_ARRAYS_IN_CASE_OF_WRONG_COUNT() macro just in case the
Count function returns a smaller value than Flatten ends up
trying to write into the array, and related to this I
got rid of the various --(flat->count) lines, since that var now
represents the allocated size of the array, up until right after
the Flatten function when it's reset to represent the data written
to the array like it used to be.

Signed-off-by: Mark Allen <markalle@us.ibm.com>