[from mpich] porting an old fix we made in OMPI's copy of romio back to mpich
Mpich PR 4948 has a commit titled
porting an old fix we made in OMPI's copy of romio back to mpich

We already have most of that commit in OMPI, but the mpich version had
a small initialization of a pointer to NULL as well, so I took that part.

There was a problem with the Alltoallv in ADIOI_GPFS_Calc_others_req().
It was using sendbuf and sdisps[] to reference the various
my_req[x].offsets, but those are all malloced separately and could
thus be further apart than the integers in sdisps[] would allow.
So this PR copies all that data into a new send buf before the
Alltoallv, then copies it back out of a new recv buf.

I don't have a testcase that demonstrates the failure, but I did
at least run a GPFS test that called MPI_File_write_all() and
uses the Alltoallv().

Signed-off-by: Mark Allen <markalle@us.ibm.com>