Remove support for C/R
Remove the base OPAL support for C/R
 - frameworks: opal/mca/crs, ompi/mca/crcp
 - components: ompi/mca/pml/crcpw
 - tools: opal-checkpoint and opal-restart.

Remove the FT event from all components and bump the MCA version of
components and modules that lost the ft_event.

Clean all the platform files and remove all references to non-existent
frameworks (mostly crs, filem and snapc).

This is scheduled for OMPI 5.0, we don't need to maintain backward
compatibility for our internal MCA with prior versions, so the
approach implemented here should be good enough.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>