Have opal_info_get return opal_cstring_t object
The key and info strings are now stored as opal_cstring_t objects so
returning them works by incrementing their reference count. The
caller is responsible for releasing the objects eventually. This avoids
many unnecessary string copy operations, which are now only required if
the string is to be modified (e.g., passed to strtok).

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>