AVX2 and AVX512* off by default to prevent frequency scaling
As discussed in #8334 the intensive use of AVX2 and AVX512 will lead to a
significant frequency scaling, with drastic impact on application performance.
Until a better way to prevent this come up, we decided to artificially disable
support for AVX2 and AVX512.

The support for AVX2 and AVX512 can be enabled by users via the corresponding
MCA parameter (op_avx_support).

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>