common_ompio_file_set_view: fix handling of  MPI_DISPLACEMENT_CURRENT
If MPI_MODE_SEQUENTIAL was used when opening the file, the special displacement  MPI_DISPLACEMENT_CURRENT
has to be used during file_set_view. The displacement is set to the current position of the shared file pointer
in this case. It is illegal to use MPI_DISPLACEMENT_CURRENT unless amode MPI_MODE_SEQUENTIAL was used.

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>