autogen: use newer config.sub|guess if available
Per https://github.com/open-mpi/ompi/issues/8410, have autogen.pl
check each config.sub and config.guess that it finds with a known-good
version of that file if the known-good version has a timestamp version
that is newer than what Autoconf installed.

We also skip updating anything in the 3rd-party tree; we don't really
want to mess with those packages.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>