mtl/ofi: Update initialization to add FI_HMEM, FI_MR_HMEM, and FI_MR_LOCAL support
Add a check to see if Libfabric has at least one provider with FI_HMEM
support, use this info to set whether or not Libfabric has CUDA support.

Add provider hints for FI_MR_LOCAL, and if Libfabric has CUDA support,
also add hints for FI_HMEM and FI_MR_HMEM.

In the case where Open MPI is built with CUDA support but Libfabric is
not, the MTL/OFI is not picked.

Signed-off-by: William Zhang <wilzhang@amazon.com>