osc/rdma: fix amo-based accumulate
This fixes an issue in osc/rdma when AMOs are used for accumulate operations
(vs get-accumulate). In this case a temporary buffer is needed to hold the
result of the operation (since fetching atomics are used). This buffer was
1) zero-sized, and 2) not freed. Both of these issues are fixed in this
commit. There was also an issue with unpacking due to using an uninitialized
convertor. The convertor is no longer passed in when no result is required.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>