btl/sm: do not set always callback frag when there is none
This fixes an issue with the btl_send function in btl/sm. The
pml/ob1 usage always provides a callback function which masked
this bug. The btl_send function was unconditionally setting
the MCA_BTL_DES_SEND_ALWAYS_CALLBACK flag for any outgoing
fragment. This is a violation of the interface which allows
the caller to leave the callback function unset. This commit
fixes the issue by protecting the code which set the flag to
first check if the callback function is non-NULL.

This commit also ensures that when a fragment is sent that
the complete flag is unset. This flag is used to return a
fragment to the sender and can confuse the receive logic
if it is set at the wrong time. This change allows the
owner of the frag to send multiple times without needing
to allocate a new frag using btl_alloc.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>