Fix core dump when performing misaligned atomics.
c_reqops MTT test has a weird alignment when creating the window,
which is perfectly valid. This will cause a segv
when performing atomic operations. While this was found on
Power, it probably is not limited to it.

Detect a misalignment, and fallback to a lock + copy method.

If it's detected that the window alignment doesn't match
the type, abandon the atomics and use a lock implementation
for all atomic operations in that window.

Signed-off-by: Austen Lauria <awlauria@us.ibm.com>