opal/hwloc: Rework shmem 'userdata' to be usable.
The 'userdata' portion of the hwloc objects is read only shmem,
so it currently isn't usable. This prevents some functionality
of this code to be lost, for example the cpu filtering.

Wrap the hwloc objects with an internal userdata and store them
in a an opal hashtable for lookup.

Minor cleanup:
- Close the shmem file.
- Finalize the hwloc component in MPI_Finalize().

Signed-off-by: Austen Lauria <awlauria@us.ibm.com>