io/ompio and fbtl: introduce a function to check for support of atomic ops
this is a new function in the fbtl framework. A default implementation is
provided in the fbtl/base directory (which simply returns false), but
an fbtl component can provide its own implementation. As of right now,
the ime and pvfs2 components use the default from base, posix provides
its own check_atomicity function.

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>