Fix case where debuggers cannot read the MPIR proctable.
Make sure the definition of the MPIR_Proctable
is in a header file that is included in the file
orted_mpir_breakpoint.c, which is compiled with -g
and compiled without optimizations.

Otherwise, the debugger (such as gdb) won't know
the complete definition of the proctable, preventing
it from being able to read it.

Since the MPIR_proctable should be accessed from
orted_submit.c and orted_mpir_breakpoint.c, move it
to the mpir_orted.h header file.

See issue: https://github.com/open-mpi/ompi/issues/8563

Signed-off-by: Austen Lauria <awlauria@us.ibm.com>