osc/ucx: convert ATOMIC_SIZE_SUPPORTED from a macro to static inline function
Generally speaking static inline functions are the better choice. Open MPI
generally leave the old macro usage in place but this macro is not ideal
as written. (_remote_addr) & 0x3 is interpreted to be a cast by
clang-format. While not right it does not have enough information in the
limited scope to decide it definitely is not a cast. I will try to patch
this case in llvm but for not it is better to just modernize this piece of
code to avoid the issue altogether.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>