decide CONVERTOR_CUDA flag based on true_lb offset into userbuf
There's a cuda check in the converter that checks one address and
caches a flag for CONVERTOR_CUDA.  But it was just checking
pUserBuf which isn't necessarily anywhere near the data, and
doesn't even need to be a legal address for that matter.  MPI
datatype displacements could put the actual data anywhere.

Since opal_cuda_check_one_buf() was already using a convertor, I
left its buffer meaning as userbuf and inside it uses the convertor's
datatype for the offset.  Then for opal_cuda_check_bufs() since it's
just passing in two addresses with no convertor/datatype information,
it's expected to handle the offset itself and pass in the final address
to check

Signed-off-by: Mark Allen <markalle@us.ibm.com>